auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic[email]==2.5.0
sqlmodel>=0.0.22
sqlalchemy>=2.0.23
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
